Linked Clone Support
This PR adds linked clone support for the vSphere provisioner.

In order to use linked clones, the template must be a VM and not
marked as a template, so when we upload the OFV and deploy it,
if UseLinkedClones is set to true, we will not mark it as a template
but instead create a snapshot which will later be referenced when
cloneTemplate is called.